MAINT: make update_subtraits a helper rather than a baseobject method
